requirements.txt
binance-connector
pandas
numpy
requests
APScheduler
python-dotenv
PyYAML
